restructured the way the project is styled to follow the Atomic Design principals
