+++ src/components/Atoms/Buttons/HeaderButton/HeaderButton.css
@@ -1,0 +1,5 @@
+.buttonContainer {
+    background: #e2cdfa;
+    border-radius: 10;
+    margin-bottom: 50px;
+}
