--- src/components/Atoms/Buttons/HomeBookButton.js
@@ -1,10 +1,0 @@
-import { style } from "../../../Style/Style.js"
-import { Button } from "@mui/material";
-
-export default function HomeBookButton() {
-  return (
-    <div style={style.buttonContainer}>
-      <Button href="/connect" style={style.bookButton}>Book An Appointment</Button>
-    </div>
-  );
-}
