+++ src/components/Atoms/Buttons/NavButtons/NavAbout.js
@@ -1,3 +1,5 @@
+import "./NavButtons.css";
+
 export default function NavBarButton() {
   return (
     <a className="navButton" href="/about">
