+++ src/components/Atoms/Buttons/NavButtons/NavButtons.css
@@ -1,0 +1,7 @@
+.navButton {
+    text-decoration: none;
+    color: #99a9b5;
+    font-family: 'Urbanist',
+        sans-serif;
+    font-size: 150%;
+}
