+++ src/components/Atoms/Buttons/NavButtons/NavConnect.js
@@ -1,3 +1,5 @@
+import "./NavButtons.css";
+
 export default function NavBarButton() {
   return (
     <a className="navButton" href="/connect">
