+++ src/components/Atoms/Buttons/NavButtons/NavHome.js
@@ -1,7 +1,9 @@
+import "./NavButtons.css";
+
 export default function NavBarButton() {
   return (
     <a className="navButton" href="/">
-        Home
+      Home
     </a>
   );
 }
