+++ src/components/Atoms/Pictures/Pictures.css
@@ -1,0 +1,4 @@
+.logoImage {
+    width: 100%;
+    height: auto;
+}
