--- src/components/Atoms/Pictures/heroHome.js
@@ -1,7 +1,0 @@
-import HeroHome from "../../../Assets/Pictures/heroHome2.jpg";
-
-export default function heroHome() {
-   return (
-     <img alt="colorful powder explosion of neon blue and pink powder against a black background" src={HeroHome} />
-   )
- }
