+++ src/components/Molecules/NavBar/NavBar.css
@@ -16,14 +16,6 @@
     width: 85vw;
 }
 
-.navButton {
-    text-decoration: none;
-    color: #99a9b5;
-    font-family: 'Urbanist',
-        sans-serif;
-    font-size: 150%;
-}
-
 /* Changes Nav Bar Font Size to 100% at >600px */
 @media screen and (max-width: 600px) {
     .navButton {
