+++ src/components/Organisms/HomePage/Header/Header.css
@@ -1,0 +1,12 @@
+.heroContainer {
+    background-image: url("../../../../Assets/Pictures/heroHome2.jpg");
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    width: 100%;
+    height: 90vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+}
