+++ src/components/Organisms/HomePage/Testimonials/Testimonials.css
@@ -1,0 +1,5 @@
+.testimonialContainer {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+}
