+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import "./Style/App.css";
+import "./App.css";
 import App from './App';
 import { BrowserRouter } from "react-router-dom";
