+++ src/pages/Homepage.js
@@ -1,5 +1,5 @@
-import HomeHeader from "../components/Organisms/HomePage/Header";
-import Testimonials from "../components/Organisms/HomePage/Testimonials";
+import HomeHeader from "../components/Organisms/HomePage/Header/Header";
+import Testimonials from "../components/Organisms/HomePage/Testimonials/Testimonials";
 import NavBar from "../components/Molecules/NavBar/NavBar";
 
 export default function Homepage() {
